[DEL] CollectionView fixed
