+++ bonheur/Sources/ViewControllers/HomeView/HomeViewController+DataSource.swift
@@ -18,4 +18,8 @@ extension HomeViewController: UICollectionViewDataSource {
         
         return cell
     }
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+
+    }
 }
